+++ restful_api_sql/config/db.config.js
@@ -2,7 +2,7 @@ module.exports = {
   HOST: "localhost",
   USER: "root", //root เป็นusername ในการต่อ database
   PASSWORD: "", //ถ้าใช้ XAMPP ก็ไม่ต้องมี PASSWORD เช่น PASSWORD: "" //ถ้าใช้ แอปเซิร์ฟ(AppServ) ก็จะมี PASSWORD เช่น PASSWORD: "12345678"
-  DATABASE: "restaurantss", //เป็นการตั้ง DATABASE ให้มีชื่อว่า restaurants
+  DB: "restaurantss", //เป็นการตั้ง DATABASE ให้มีชื่อว่า restaurants
 };
 
 //config คือเป็นการตั้งค่าต่างๆ
