+++ restful_api_sql/models/restaurant.model.js
@@ -1,1 +1,35 @@
-const {DataType} = require
+const { DataTypes } = require("sequelize");
+const sequelize = require("./db");
+
+//Define the restaurant
+const Restaurant = sequelize.define("restaurant", {
+  id: {
+    type: DataTypes.INTEGER,
+    primaryKey: true,
+    autoIncrement: true,
+  },
+  name: {
+    type: DataTypes.STRING,
+    allowNull: false,
+  },
+  type: {
+    type: DataTypes.STRING,
+    allowNull: false,
+  },
+  imageURL: {
+    type: DataTypes.STRING,
+    allowNull: false,
+  },
+  createdAt: {
+    type: DataTypes.DATE,
+    allowNull: true,
+    defaultValue: DataTypes.NOW,
+  },
+  updatedAt: {
+    type: DataTypes.DATE,
+    allowNull: true,
+    defaultValue: DataTypes.NOW,
+  },
+});
+
+module.exports = Restaurant;
